appengine/requirements.txt
Flask
Pillow
requests
google-auth
google-cloud-storage